feat: insert child machine trees into parent tree (#2)
Any components rendered by a child machine are now inserted into the component tree where the child machine was invoked. This allows for much more natural composition of statecharts and component rendering.

BREAKING CHANGE: The output format has changed to no longer have an array of machines at the top-level, instead it is just the top-level components representing active states. Invoked machines are no longer part of the top-level array but now inserted into the tree of components based on the location of the component that invoked them.