feat: make second callback arg useful (#42)
- Documenting new features
- Rearranging tests a bit to take advantage of new features

BREAKING CHANGE: previously the second arg to the callback function had a single `data` property on it representing the last `State` object seen by the top-level machine. Now it has `state` (same as data was previously), and some bound APIs for interacting with the statechart: `.hasTag()`, `.broadcast()`, and `.matches()`. These are the same APIs available on the `ComponentTree` instance but made available through the callback args for convenience.